+++ backend/resolvers/Mutation/index.js
@@ -1,0 +1,5 @@
+const authentication = require('./Authentication')
+
+module.exports = {
+  ...authentication
+}
